Refacto de la classe de paramétrage du modèle

+++ src/main/java/ia/ecj/statistics/StatisticsSaveModel.java
@@ -34,7 +34,7 @@ public class StatisticsSaveModel implements CTF_CMAES_StatListener {
             if (model.isTerritoryCompass()) {
                 writerCTF.write("ia.perception.TerritoryCompass;ALLY;NEAREST\n");
             }
-            for (List<Integer> raycast : model.getRaycasts()) {
+            for (List<Double> raycast : model.getRaycasts()) {
                 writerCTF.write("ia.perception.PerceptionRaycast");
                 for (int i = 0; i < raycast.size(); i++) {
                     writerCTF.write(";" + raycast.get(i));
